add JobSystem with SPSC ring and worker threads
Adds a global thread-safe job system backed by per-worker SpscRing<Job>
queues. Workers wait on condition variables and drain their rings without
locking. An overflow ring on JobSystem handles backpressure when worker
rings are full. Lifecycle is RAII. Includes 14 tests covering SpscRing
correctness, dispatch from multiple threads, shutdown, and overflow.

+++ tests/list.test.cpp
@@ -575,7 +575,7 @@ DTEST(moveAssignmentHeapToHeap) {
 
   ASSERT_EQ(list2.getSize(), 100);
   for (int i = 0; i < 100; ++i) {
-    ASSERT_EQ(list2[i], i);
+    ASSERT_EQ(list2[static_cast<u64>(i)], i);
   }
 }
 
@@ -608,6 +608,6 @@ DTEST(copyAssignmentHeapToHeap) {
 
   ASSERT_EQ(list2.getSize(), 100);
   for (int i = 0; i < 100; ++i) {
-    ASSERT_EQ(list2[i], i);
+    ASSERT_EQ(list2[static_cast<u64>(i)], i);
   }
 }
